refactoring

+++ raspberry-pi-client/c/rc-car.c
@@ -25,12 +25,12 @@ void move(RcCar *self, const int *speed, const char *direction) {
     int pulseWidth = CAR_ESC_NEUTRAL_MAX_PWM;
 
     if (strcmp(direction, "forward") == 0) {
-        pulseWidth = (int)floor(
+        pulseWidth = (int)floorf(
             CAR_ESC_NEUTRAL_PWM +
             ((float)(*speed) / 100.0f) * (CAR_ESC_NEUTRAL_MAX_PWM - CAR_ESC_NEUTRAL_PWM)
         );
     } else if (strcmp(direction, "backward") == 0) {
-        pulseWidth = (int)floor(
+        pulseWidth = (int)floorf(
             CAR_ESC_NEUTRAL_PWM +
             ((float)(*speed) / 100.0f) * (CAR_ESC_NEUTRAL_PWM - CAR_ESC_NEUTRAL_MIN_PWM)
         );
